pyproject.toml
[tool.poetry]
name = "covid19-vaccine-tracker-india"
version = "0.1.0"
description = ""
authors = ["Sidharth Arya <sidhartharya10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.7"
pandas = "^1.3.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
